e2e: drop vm-force-restart()
Drop the vm-force-restart() function and just use vm-reboot() instead.
Improve the vm-reboot() to do a forced restart (via Qemu command) if the
normal reboot fails - something that the old vm-force-restart() tried to
do. However, the old vm-force-restart() was broken as we always
unconditionally forced a qemu system resstart independent of whether the
normal shutdown succeeded or not. This led to races in rare cases where
the qemu command was executed precisely when the container was in not
running state. Not to mention that this practice never resulted in
graceful restarts.

Tha patch increases the sleep after reboot command to 10 seconds to
match what vm-force-restart() had, apparently to avoid inadvertent ssh
logins to a shutting down system (before ssh daemon is killed and actual
reboot happens).